% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strptimer.R
\name{base_options}
\alias{base_options}
\title{Get a list of optional modifications to a unit of time}
\usage{
base_options(specific_base)
}
\arguments{
\item{specific_base}{A base unit of time to explore (see vignette)}
}
\description{
See vignette
}

