% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strptimer.R
\name{time_format_}
\alias{time_format_}
\title{Standard evaluation version of time_format}
\usage{
time_format_(args)
}
\arguments{
\item{args}{A list of unevaluated time format pieces}
}
\description{
See vignette for lazyeval
}

